dev/6.greengrass_core/edge-components/incident-message-forwarder/requirements.txt
awsiotsdk==1.11.9
